modules/llm_client.py
import os
from dotenv import load_dotenv
from huggingface_hub import InferenceClient

load_dotenv(dotenv_path=os.path.join(os.path.dirname(__file__), "..", ".env"), override=True)

MODEL = "Qwen/Qwen2.5-7B-Instruct"

_client = None

def get_client() -> InferenceClient:
    global _client
    if _client is None:
        token = os.environ.get("HF_TOKEN")
        if not token:
            raise ValueError("HF_TOKEN not found in .env file")
        _client = InferenceClient(token=token)
    return _client

def ask(system_prompt: str, user_prompt: str, max_tokens: int = 2048) -> str:
    client = get_client()
    response = client.chat_completion(
        model=MODEL,
        messages=[
            {"role": "system", "content": system_prompt},
            {"role": "user", "content": user_prompt},
        ],
        max_tokens=max_tokens,
        temperature=0.2,
    )
    return response.choices[0].message.content
