requirements.txt
google-generativeai
streamlit
chromadb
sentence-transformers
pdfplumber
python-docx
python-dotenv